fix QH_DP macro

+++ QHCoreLib/UI/QHUIDefines.h
@@ -17,11 +17,11 @@
 #define QH_SCREEN_PORTRAIT_WIDTH    (MIN(QH_SCREEN_WIDTH, QH_SCREEN_HEIGHT))
 #define QH_SCREEN_PORTRAIT_HEIGHT   (MAX(QH_SCREEN_WIDTH, QH_SCREEN_HEIGHT))
 
-#define QH__DP(size, _base_screen_width) ((size) / (_base_screen_width) * QH_SCREEN_PORTRAIT_WIDTH)
+#define QH__DP(size, _base_screen_width) ((CGFloat)(size) / (_base_screen_width) * QH_SCREEN_PORTRAIT_WIDTH)
 
-#define QH_DP_320(size)     QH__DP(size, 320)
-#define QH_DP_375(size)     QH__DP(size, 375)
-#define QH_DP_414(size)     QH__DP(size, 414)
+#define QH_DP_320(size)     QH__DP(size, 320.0)
+#define QH_DP_375(size)     QH__DP(size, 375.0)
+#define QH_DP_414(size)     QH__DP(size, 414.0)
 
 #define QH_DP(size) QH_DP_375(size)
